refactor: better name for magic sign in flow
