--- MyFinance-Backend/MyFinance.Application/UseCases/Users/Commands/CreateMagicSignInToken/CreateMagicSignInTokenCommand.cs
@@ -1,7 +1,0 @@
-﻿using MyFinance.Application.Abstractions.RequestHandling.Commands;
-
-namespace MyFinance.Application.UseCases.Users.Commands.CreateMagicSignInToken;
-
-public sealed record CreateMagicSignInTokenCommand(string Email) : ICommand
-{
-}
