+++ MyFinance-Backend/MyFinance.Application/UseCases/Users/Commands/SendMagicSignInEmail/SendMagicSignInEmailCommand.cs
@@ -1,0 +1,7 @@
+﻿using MyFinance.Application.Abstractions.RequestHandling.Commands;
+
+namespace MyFinance.Application.UseCases.Users.Commands.SendMagicSignInEmail;
+
+public sealed record SendMagicSignInEmailCommand(string Email) : ICommand
+{
+}
