+++ MyFinance-Backend/MyFinance.Application/UseCases/Users/Commands/SendMagicSignInEmail/SendMagicSignInEmailHandler.cs
@@ -3,17 +3,17 @@ using MyFinance.Application.Abstractions.Persistence.Repositories;
 using MyFinance.Application.Abstractions.RequestHandling.Commands;
 using MyFinance.Application.Abstractions.Services;
 
-namespace MyFinance.Application.UseCases.Users.Commands.CreateMagicSignInToken;
+namespace MyFinance.Application.UseCases.Users.Commands.SendMagicSignInEmail;
 
-internal sealed class CreateMagicSignInTokenHandler(
+internal sealed class SendMagicSignInEmailHandler(
     ISignInManager signInManager,
     IUserRepository userRepository)
-    : ICommandHandler<CreateMagicSignInTokenCommand>
+    : ICommandHandler<SendMagicSignInEmailCommand>
 {
     private readonly ISignInManager _signInManager = signInManager;
     private readonly IUserRepository _userRepository = userRepository;
 
-    public async Task<Result> Handle(CreateMagicSignInTokenCommand command, CancellationToken cancellationToken)
+    public async Task<Result> Handle(SendMagicSignInEmailCommand command, CancellationToken cancellationToken)
     {
         var user = await _userRepository.GetByEmailAsync(command.Email, cancellationToken);
