--- MyFinance-Backend/MyFinance.Contracts/User/Requests/CreateMagicSignInTokenRequest.cs
@@ -1,5 +1,0 @@
-﻿namespace MyFinance.Contracts.User.Requests;
-
-public sealed record CreateMagicSignInTokenRequest(string Email)
-{
-}
