+++ MyFinance-Backend/MyFinance.Contracts/User/Requests/SendMagicSignInEmailRequest.cs
@@ -1,0 +1,5 @@
+﻿namespace MyFinance.Contracts.User.Requests;
+
+public sealed record SendMagicSignInEmailRequest(string Email)
+{
+}
